destroy-data.sh: Ask the user for confirmation add all volumes
As this script is destructive, ask the user for confirmation first.
Also remove the second kernelci volume.

+++ destroy-data.sh
@@ -1,4 +1,15 @@
 #!/bin/bash
 
+if [ "$1" != "-y" ];then
+    while true; do
+        read -p "CAUTION: This is a destructive script that will destroy all kernel-ci docker data. Proceed ? " yn
+        case $yn in
+            [Yy]* ) break;;
+            [Nn]* ) echo "Nothing done";exit;;
+            * ) echo "Please answer yes(y/Y) or no(n/N).";;
+        esac
+    done
+fi
 echo "-> Remove volume..."
 docker volume rm kernelci_data
+docker volume rm kernelci_kci
